pyproject.toml
[tool.poetry]
name = "y2zim"
version = "0.1.0"
description = "A tool for creating a subset off a YouTube channel or playlist(s)"
authors = ["Blondel MONDESIR <blondel.md@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[overrides]
iso-639 = "0.4.5"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
yt-dlp = "^2023.3.4"
rich = "^13.3.3"
python-slugify = "^8.0.1"
pytube = "^12.1.3"
python-dateutil = "^2.8.2"
jinja2 = "^3.0.0"
markupsafe = "^2.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"


[tool.poetry.group.fmt.dependencies]
black = "^23.3.0"
isort = "^5.12.0"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.4"
mkdocs-htmlproofer-plugin = "^0.11.0"
mkdocstrings = "^0.20.0"
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.6.0"


[tool.poetry.group.nox.dependencies]
nox-poetry = "^1.0.2"


[tool.poetry.group.type-check.dependencies]
mypy = "^1.1.1"
types-requests = "^2.28.11.17"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.23"
flake8-comprehensions = "^3.11.1"
pep8-naming = "^0.13.3"
flake8-pyproject = "^1.2.3"

[tool.poetry.scripts]
y2zim = "y2zim.entrypoint:main"

[tool.mypy]
ignore_missing_imports = true
strict = true
# If certain strict config options are too pedantic for a project,
# disable them selectively here by setting to false.

# Note: This configuration is suppported by flake8-pyproject.
[tool.flake8]
# Use the more relaxed max line length permitted in PEP 8.
max-line-length = 99
# This ignore is required by black.
extend-ignore = ["E203"]
# flake8-bugbear config.

[tool.black]
line-length = 99
target-version = ["py38", "py39", "py310"]
# black will automatically exclude all files listed in .gitignore
# If you need to exclude additional folders, consider using extend-exclude to avoid disabling the
# default .gitignore behaviour.

[tool.isort]
profile = "black"
line_length = 99
force_sort_within_sections = true
# Inform isort of paths to import names that should be considered part of the "First Party" group.
src_paths = ["src/y2zim", "tests"]
skip_gitignore = true
# If you need to skip/exclude folders, consider using skip_glob as that will allow the
# isort defaults for skip to remain without the need to duplicate them.

[tool.pytest.ini_options]
addopts = [
    "--strict-config",
    "--strict-markers",
]
xfail_strict = true
filterwarnings = [
    # When running tests, treat warnings as errors (e.g. -Werror).
    # See: https://docs.pytest.org/en/latest/reference/reference.html#confval-filterwarnings
    "error",
    # Add additional warning supressions as needed here. For example, if a third-party library
    # is throwing a deprecation warning that needs to be fixed upstream:
    # "ignore::DeprecationWarning:typer",
]

[tool.coverage.run]
branch = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

